chore: consolidate files
